build first draft for index screen
