+++ frontend/src/components/card/card.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { styled } from "styled-components";
 
 import { CardProps } from "./card.props";
-import { shadow, spacing } from "../../theme";
+import { color, fontSize, fontWeight, shadow, spacing } from "../../theme";
 
 const StyledCard = styled.div`
   background: white;
@@ -20,8 +20,10 @@ const CardTitle = styled.div`
   justify-content: space-between;
   align-items: center;
   width: 100%;
-  font-size: 18px;
-  font-weight: bold;
+  font-size: ${fontSize("h3")};
+  line-height: ${fontSize("h3")};
+  font-weight: ${fontWeight("bold")};
+  color: ${color("primary")};
   padding: ${spacing("small")};
 `;
